Update to use new format for USAA output
USAA updated their UI and that appears to update the output format
of the CSV file that we export.

The field ordering and headers appear to be different, and
the first line needs to be skipped during import.

I updated the field names, and added a feature where an aribitrary
number of lines can be skipped during import, I've set that to 1 for
now.